chore: extract browserInterface
