+++ kernel/packages/shared/renderer-interface/browserInterface/browserInterfaceType.ts
@@ -14,6 +14,8 @@ export type browserInterfaceType = {
   TermsOfServiceResponse: (sceneId: string, accepted: boolean, dontShowAgain: boolean) => void
   MotdConfirmClicked: () => void
   GoTo: (data: { x: number; y: number }) => void
+  GoToCrowd: () => void
+  GoToMagic: () => void
   LogOut: () => void
   SaveUserAvatar: (changes: { face: string; body: string; avatar: Avatar }) => void
   SaveUserTutorialStep: (data: { tutorialStep: number }) => void
